test(api): keep admin shim coverage out of async legacy file
